Fix stray keyboard input on focus gain
When the application window receives focus while some keys are held
down, SDL will generate key-up events for all the keys that were still
down when the focus was lost. Presumably, this is done so that SDL
doesn't need to track the state of those keys while the game window
isn't focused (which may not even be possible).

The converse is that when the game window is once again focused, SDL
respectively emits key-down events for any keys that are held down
when the focus is gained. This causes the game to think that the
player pressed some keys intended for it, however these keypresses
were actually meant for the window manager or another
application (esp. when combined with modifier keys).

We can detect these generated key presses my taking advantage of the
"timestamp" field in SDL event structures, and comparing the
timestamps of received keyboard events with that of the most recent
focus-in event.

This commit thus fixes the above-described problem by filtering out
keyboard events received less than 10 milliseconds since the last
focus event.